# docs/swagger.yaml
definitions:
...
  handler.StockLimitReq:
    properties:
      maxStock:
        type: number
      minStock:
        type: number
      productIds:
        items:
          type: integer
        type: array
    required:
    - productIds
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /stocks/status:
    get:
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      - description: 商品名称/编号
        in: query
        name: keyword
        type: string
      - description: 仓库ID
        in: query
        name: warehouseId
        type: integer
      - description: 商品分类ID（含子分类）
        in: query
        name: categoryId
        type: integer
      - description: 库存状态 in/out/belowMin/aboveMax
        in: query
        name: stockStatus
        type: string
      - description: 商品状态 1启用 0停用
        in: query
        name: productStatus
        type: string
      - description: 仅显示有账面库存商品
        in: query
        name: onlyWithStock
        type: boolean
      - description: product 按商品 / category 按分类
        in: query
        name: groupBy
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 库存状况表（按商品/按分类）
      tags:
      - 库存
  /stocks/status/{productId}/flows:
    get:
      parameters:
      - description: 商品ID
        in: path
        name: productId
        required: true
        type: integer
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 商品出入库明细
      tags:
      - 库存
  /stocks/status/{productId}/warehouses:
    get:
      parameters:
      - description: 商品ID
        in: path
        name: productId
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 商品库存分布
      tags:
      - 库存
  /stocks/status/limits:
    put:
      parameters:
      - description: 商品ID列表与上下限
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.StockLimitReq'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 批量设置商品库存上下限
      tags:
      - 库存
swagger: "2.0"